ATM cringe fixed

+++ sem5/idea/ATM/src/Main.java
@@ -4,7 +4,7 @@ public class Main {
     public static void main(String[] args) {
         System.out.println("Enter count bills of 20, 50 and 100");
         Scanner scanner = new Scanner(System.in);
-        ATM atm = new ATM(new CashSet(scanner.nextInt(), scanner.nextInt(), scanner.nextInt()));
+        ATM atm = new ATM(new CashSet(scanner.nextInt(), scanner.nextLong(), scanner.nextLong()));
         System.out.println("Enter amount you need to get:");
         System.out.println( atm.takeCash(scanner.nextInt()) );
         System.out.println(atm.getStorage());
